+++ src/theme/Navbar/hooks/useKeyboardShortcut.js
@@ -1,0 +1,14 @@
+import { useEffect } from "react";
+
+export default function useKeyboardShortcut(key, callback) {
+  useEffect(() => {
+    const handleKeyDown = (event) => {
+      if ((event.metaKey || event.ctrlKey) && event.key.toLowerCase() === key) {
+        callback();
+        event.preventDefault();
+      }
+    };
+    window.addEventListener("keydown", handleKeyDown);
+    return () => window.removeEventListener("keydown", handleKeyDown);
+  }, [key, callback]);
+}
