+++ src/theme/Navbar/styles/navbar.module.scss
@@ -1,0 +1,172 @@
+/* Define common values as CSS custom properties */
+:root {
+  --secondary-navbar-height: 60px;
+  --docs-divider-color: #ccc;
+}
+
+/* NAVBAR STYLES */
+.defaultNavbar {
+  position: sticky;
+  z-index: 1000;
+}
+
+.onlyDefaultNavbar {
+  top: 0 !important;
+}
+
+.secondaryNavbar {
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  background-color: var(--docs-sidebar-bg);
+  padding: 0 0 0 10px;
+  border-top: 1px solid var(--ifm-toc-border-color);
+  border-bottom: 1px solid var(--ifm-toc-border-color);
+  position: sticky;
+  width: 100%;
+  height: var(--secondary-navbar-height);
+  z-index: 1001;
+}
+
+/* Placeholder to prevent content shift when secondaryNavbar becomes fixed */
+.secondaryNavbarPlaceholder {
+  height: var(--secondary-navbar-height);
+  display: none;
+}
+
+.secondaryNavbar.fixOnTop {
+  position: fixed;
+}
+
+.secondaryNavbar.fixOnTop + .secondaryNavbarPlaceholder {
+  display: block;
+}
+
+/* NAVIGATION LINKS */
+.secondaryNavLinks {
+  display: flex;
+  gap: 40px;
+  align-items: center;
+  height: 100%;
+  margin-left: 20px; // Shift items to the right
+}
+
+.navLink {
+  color: var(--ifm-navbar-link-color);
+  padding: 10px 0;
+  display: flex;
+  align-items: center;
+  height: 100%;
+}
+
+.navLink:hover {
+  text-decoration: none;
+  color: var(--ifm-navbar-hover-link-color);
+}
+
+.activeNavLink {
+  color: var(--ifm-navbar-link-active-color) !important;
+  border-bottom: 3px solid var(--ifm-navbar-link-active-color);
+}
+
+/* Container for the dropdown */
+.navDropdown {
+  position: relative;
+  display: inline-block;
+}
+
+/* Dropdown toggle styling */
+.navLinkWithArrow {
+  display: inline-flex; // Ensure text and icon are on the same line
+  align-items: center; // Vertically center them
+  cursor: pointer;
+}
+
+.navLinkWithArrow i {
+  font-size: 12px;
+  margin-left: 4px; // Spacing between text and icon
+  display: inline-block; // Make sure the transform applies
+}
+
+/* Container for dropdown items */
+.dropdownContent {
+  display: none;
+  position: absolute;
+  background-color: var(--docs-modal-bg);
+  min-width: 160px;
+  box-shadow: 0px 8px 16px rgba(0, 0, 0, 0.2);
+  border-radius: 8px;
+  z-index: 1;
+}
+
+/* Show dropdown on hover */
+.navDropdown:hover .dropdownContent {
+  display: block;
+}
+
+/* Style individual dropdown items */
+.dropdownItem {
+  color: var(--ifm-navbar-link-color);
+  padding: 12px 16px;
+  text-decoration: none;
+  display: block;
+}
+
+.dropdownItem:hover {
+  color: var(--ifm-navbar-link-color) !important;
+}
+
+/* Override active styling for dropdown items */
+.dropdownItem.activeNavLink {
+  border-bottom: none !important;
+  color: var(--ifm-navbar-link-color) !important;
+  background: none !important;
+}
+
+.dropdownDivider {
+  height: 1px; // Set the thickness of the divider
+  background-color: var(--ifm-toc-border-color); // Use a light gray
+  width: 100%; // Make it span the full width of the dropdown
+  margin: 0; // Remove any default margin
+}
+
+/* RESPONSIVE STYLES FOR NAVBAR */
+@media (max-width: 999px) {
+  .secondaryNavbar {
+    /* Always show the horizontal scrollbar */
+    overflow-x: scroll;
+    white-space: nowrap;
+    -webkit-overflow-scrolling: touch;
+    padding-bottom: 5px;
+
+    /* Optional: style the scrollbar for Firefox */
+    scrollbar-width: thin;
+    scrollbar-color: var(--ifm-navbar-link-active-color);
+    scrollbar-gutter: stable; /* Reserves space for the scrollbar */
+  }
+
+  /* Optional: style the scrollbar for WebKit-based browsers */
+  .secondaryNavbar::-webkit-scrollbar {
+    height: 6px; /* Adjust the scrollbar height as needed */
+  }
+  .secondaryNavbar::-webkit-scrollbar-track {
+    background: transparent;
+  }
+  .secondaryNavbar::-webkit-scrollbar-thumb {
+    background-color: var(--ifm-navbar-link-active-color);
+    border-radius: 3px;
+  }
+}
+
+@media (max-width: 599px) {
+  .secondaryNavLinks {
+    display: flex;
+    flex-wrap: nowrap;
+    gap: 20px;
+  }
+
+  .navLink {
+    flex-shrink: 0;
+    padding: 10px;
+  }
+}
